Update staking and balance endpoints

+++ internal/service/address/entity/balance.go
@@ -1,7 +1,7 @@
 package entity
 
 type Balance struct {
-	Address           string  `json:"address"`
-	Balance           float64 `json:"balance"`
-	ColdStakedBalance float64 `json:"coldStakedBalance"`
+	Address           string `json:"address"`
+	Balance           int64  `json:"balance"`
+	ColdStakedBalance int64  `json:"coldStakedBalance"`
 }
